+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - component `SelectWithLogo`
 - component `ValidMultiCreatableFieldText`
+- `async` prop for `Select` component
+- refactor `ValidSelect` & `ValidSelectAsync`
 
 ## [0.10.3]
 ### Added